// packages/gatsby-theme-vidyaa/src/types/type-trans-links.ts
export interface ItransLinks {
  linkReviews: {
    url: string
    trans: string
  }
  linkPictures: {
    url: string
    trans: string
  }
  linkVideos: {
    url: string
    trans: string
  }
  linkBooking: {
    url: string
    trans: string
  }
  linkFacebook: {
    url: string
    trans: string
  }
  linkInstagram: {
    url: string
    trans: string
  }
  linkEmail: {
    url: string
    trans: string
  }
  linkMap: {
    url: string
    trans: string
  }
}